services/order_service/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
mysqlclient==2.1.1
PyJWT==2.6.0
django-cors-headers==4.1.0
python-dotenv==1.0.0
pika==1.3.2
requests==2.31.0  # Add this line
python-decouple
